Separated SegmentCursor to its own file
