+++ drools-core/src/main/java/org/drools/core/phreak/PhreakJoinNode.java
@@ -48,8 +48,8 @@ public class PhreakJoinNode {
                        LeftTupleSink sink,
                        BetaMemory bm,
                        TupleSets srcLeftTuples,
-                       TupleSets trgLeftTuples,
-                       TupleSets stagedLeftTuples) {
+                       TupleSets stagedLeftTuples,
+                       TupleSets trgLeftTuples) {
 
         TupleSets srcRightTuples = bm.getStagedRightTuples().takeAll();
