+++ drools-metric/src/main/java/org/drools/metric/phreak/PhreakGroupByNodeMetric.java
@@ -36,13 +36,13 @@ public class PhreakGroupByNodeMetric extends PhreakGroupByNode {
                         LeftTupleSink sink,
                         AccumulateNode.AccumulateMemory am,
                         TupleSets srcLeftTuples,
-                        TupleSets trgLeftTuples,
-                        TupleSets stagedLeftTuples) {
+                        TupleSets stagedLeftTuples,
+                        TupleSets trgLeftTuples) {
 
         try {
             MetricLogUtils.getInstance().startMetrics(accNode);
 
-            super.doNode(accNode, sink, am, srcLeftTuples, trgLeftTuples, stagedLeftTuples);
+            super.doNode(accNode, sink, am, srcLeftTuples, stagedLeftTuples, trgLeftTuples);
 
         } finally {
             MetricLogUtils.getInstance().logAndEndMetrics();
