+++ drools-metric/src/main/java/org/drools/metric/phreak/PhreakQueryNodeMetric.java
@@ -33,17 +33,17 @@ public class PhreakQueryNodeMetric extends PhreakQueryNode {
     }
 
     @Override
-    public void doNode(QueryElementNode queryNode,
+    public void doNode(StackEntry stackEntry,
+                       QueryElementNode queryNode,
                        QueryElementNodeMemory qmem,
-                       StackEntry stackEntry,
                        TupleSets srcLeftTuples,
-                       TupleSets trgLeftTuples,
-                       TupleSets stagedLeftTuples) {
+                       TupleSets stagedLeftTuples,
+                       TupleSets trgLeftTuples) {
 
         try {
             MetricLogUtils.getInstance().startMetrics(queryNode);
 
-            super.doNode(queryNode, qmem, stackEntry, srcLeftTuples, trgLeftTuples, stagedLeftTuples);
+            super.doNode(stackEntry, queryNode, qmem, srcLeftTuples, stagedLeftTuples, trgLeftTuples);
 
         } finally {
             MetricLogUtils.getInstance().logAndEndMetrics();
